data/requirements.txt
openml
kafka-python
pandas
